+++ src/chart/generator/plot.h
@@ -84,6 +84,8 @@ public:
 	void detachOptions();
 	[[nodiscard]] bool isEmpty() const;
 
+	[[nodiscard]] Math::Range<> getMarkersBounds(AxisId axisId) const;
+
 	static bool dimensionMatch(const Plot &a, const Plot &b);
 	static bool hasMarkerChange(const Plot &source,
 	    const Plot &target);